# docs/openapi.yaml
openapi: 3.0.0
info:
    title: API de Partidos de Deportes en Equipo
    description: API para la gestión de usuarios, agrupaciones y partidos con sistema dinámico de méritos.
    version: 1.5.0
security:
    - bearerAuth: []
paths:
...
    /api/groups:
        get:
            summary: Listar mis grupos
            operationId: listMyGroups
            tags: [Groups]
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                type: array
                                items: { $ref: "#/components/schemas/Group" }
        post:
            summary: Crear Agrupación
            description: Crea una agrupación con configuración de méritos por defecto o personalizada. Solo Owners.
            operationId: createGroup
            tags: [Groups]
            requestBody:
                required: true
                content:
                    application/json:
                        schema: { $ref: "#/components/schemas/GroupInput" }
            responses:
                "201":
                    content:
                        application/json:
                            schema: { $ref: "#/components/schemas/Group" }
                "400": { $ref: "#/components/responses/BadRequest" }
                "409": { $ref: "#/components/responses/Conflict" }
    /api/groups/{id}:
        get:
            summary: Obtener detalle de grupo
            operationId: getGroupById
            tags: [Groups]
            parameters: [{ $ref: "#/components/parameters/Id" }]
            responses:
                "200":
                    content:
                        application/json:
                            schema: { $ref: "#/components/schemas/Group" }
                "403": { $ref: "#/components/responses/Forbidden" }
                "404": { $ref: "#/components/responses/NotFound" }
        patch:
            summary: Actualizar grupo
            description: Actualiza metadatos, estado o configuración de méritos. Solo Owners.
            operationId: updateGroup
            tags: [Groups]
            parameters: [{ $ref: "#/components/parameters/Id" }]
            requestBody:
                content:
                    application/json:
                        schema: { $ref: "#/components/schemas/GroupUpdate" }
            responses:
                "200":
                    content:
                        application/json:
                            schema: { $ref: "#/components/schemas/Group" }
                "403": { $ref: "#/components/responses/Forbidden" }
    /api/groups/{id}/matches:
        get:
            summary: Listar partidos de un grupo
            description: Obtiene todos los partidos organizados por el grupo. Solo miembros.
            operationId: listGroupMatches
            tags: [Matches]
            parameters: [{ $ref: "#/components/parameters/Id" }]
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                type: array
                                items: { $ref: "#/components/schemas/Match" }
                "403": { $ref: "#/components/responses/Forbidden" }
...
components:
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
    parameters:
        Id:
            name: id
            in: path
            required: true
            schema: { type: string, format: uuid }
    schemas:
...
        Group:
            type: object
            properties:
                id: { type: string, format: uuid }
                name: { type: string }
                description: { type: string }
                isActive: { type: boolean }
                createdAt: { type: string, format: date-time }
                updatedAt: { type: string, format: date-time }
                meritConfigMaxMatches: { type: integer, default: 10 }
                meritPointsPlayed: { type: integer, default: 3 }
                meritPointsNoShow: { type: integer, default: -5 }
                meritPointsReserve: { type: integer, default: 1 }
                meritPointsPositiveAttitude: { type: integer, default: 1 }
                meritPointsNegativeAttitude: { type: integer, default: -1 }
                meritConfigHoursBeforePenalty: { type: integer, default: 12 }
                meritPointsLateCancel: { type: integer, default: -2 }
        GroupInput:
            type: object
            required: [name]
            properties:
                name: { type: string }
                description: { type: string }
                meritConfigMaxMatches: { type: integer }
                meritPointsPlayed: { type: integer }
                meritPointsNoShow: { type: integer }
                meritPointsReserve: { type: integer }
                meritPointsPositiveAttitude: { type: integer }
                meritPointsNegativeAttitude: { type: integer }
                meritConfigHoursBeforePenalty: { type: integer }
                meritPointsLateCancel: { type: integer }
        GroupUpdate:
            type: object
            properties:
                name: { type: string }
                description: { type: string }
                isActive: { type: boolean }
                meritConfigMaxMatches: { type: integer }
                meritPointsPlayed: { type: integer }
                meritPointsNoShow: { type: integer }
                meritPointsReserve: { type: integer }
                meritPointsPositiveAttitude: { type: integer }
                meritPointsNegativeAttitude: { type: integer }
                meritConfigHoursBeforePenalty: { type: integer }
                meritPointsLateCancel: { type: integer }
        RGB:
            type: object
            properties:
                r: { type: number, minimum: 0, maximum: 1 }
                g: { type: number, minimum: 0, maximum: 1 }
                b: { type: number, minimum: 0, maximum: 1 }
        SportType: { type: string, enum: [football, basketball, padel] }
        MatchStatus: { type: string, enum: [planning, playing, finished, cancelled] }
...
        Match:
            type: object
            properties:
                id: { type: string, format: uuid }
                groupId: { type: string, format: uuid }
                sport: { $ref: "#/components/schemas/SportType" }
                scheduledAt: { type: string, format: date-time }
                durationMinutes: { type: integer }
                capacity: { type: integer }
                location: { type: string }
                status: { $ref: "#/components/schemas/MatchStatus" }
                isLocked: { type: boolean }
                teamAColor: { $ref: "#/components/schemas/RGB" }
                teamBColor: { $ref: "#/components/schemas/RGB" }
                createdAt: { type: string, format: date-time }
                updatedAt: { type: string, format: date-time }
...
    responses:
        BadRequest:
            description: Error en los datos.
            content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
        Forbidden:
            description: Permisos insuficientes o recurso bloqueado.
            content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
        NotFound:
            description: Recurso no encontrado.
            content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
        Conflict:
            description: Conflicto de estado o duplicidad.
            content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...